// src/app/models/user.ts
export class User
{
    id: string;
    avatar: string;
    username: string;
    email: string;
    phone_number: string;
    password: string;
    passwordConfirm: string;
    active_status: number;
    roles: any[];
    role_id: number;
    smsbot_number: string;
    store_name: string;
    store_link: string;
}